speech-asr/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
torch
openai-whisper